tui: spin widget.rs into a folder
